Resolve "JXW-三期改动"

Closes #7

See merge request cerong_dev/mall!126

+++ mall-portal/src/main/java/com/macro/mall/portal/service/impl/XmsCustomerSkuStockServiceImpl.java
@@ -61,7 +61,7 @@ public class XmsCustomerSkuStockServiceImpl extends ServiceImpl<XmsCustomerSkuSt
             List<XmsCustomerProduct> productList = this.xmsCustomerProductMapper.selectList(productQueryWrapper);
 
             if(CollectionUtil.isNotEmpty(productList)){
-                List<Integer> collect = productList.stream().mapToInt(XmsCustomerProduct::getId).boxed().collect(Collectors.toList());
+                List<Long> collect = productList.stream().mapToLong(XmsCustomerProduct::getId).boxed().collect(Collectors.toList());
                 lambdaQuery.in(XmsCustomerSkuStock::getProductId, collect);
                 productList.clear();
             }
